everything works, execept commands: help, meeting, random-number and role
