+++ commands/utility/template
@@ -1,0 +1,10 @@
+const { SlashCommandBuilder } = require('discord.js');
+
+module.exports = {
+	data: new SlashCommandBuilder()
+		.setName('')
+		.setDescription(''),
+	async execute(interaction) {
+    // logic
+  },
+};
